move sidebar to lib/components
